+++ doc/user/content/release-notes.md
@@ -103,6 +103,7 @@ List new features before bug fixes.
 
 {{% version-header v0.9.13 %}}
 
+- Add support for `LIST(<subquery>)` constructor.
 
 {{% version-header v0.9.12 %}}